Prevent parent loggers from overriding children
The config file gets parsed into an associative array, and hence
provides no guarantees about the order in which properties are
retrieved. However, this is not suitable for configuring loggers, as
this could result in parent loggers overriding the configurations of
child loggers. To avoid this, the names of all loggers are first sorted
in ascending order of their respective lengths, and the loggers are then
configured in this order. This ensures that parent loggers are always
configured before child loggers.

Fixes #1153.