Update README and RELEASE_NOTES
Add Preview Release note to make it more explicit that there are some
thing to have in mind in this project.

Also add dependencies section and building instructions, including
instructions to convert to D2.

Finally clean the RELEASE_NOTES, adding a note saying there is not much
to see here for this release.