Update to v2.1.0
This commit has all the changes done between v2.0.8 and v2.1.0 in the
internal project. Unfortunately we are still not ready to move the whole
development to the open source project and backporting a full minor release
commit by commit is a huge task, so for now for big updates well just have
to live with one big diff commit.

Also MakD is added as a submodule, for some reason it seems like it was
forgotten in the first release.

Fixes #2 and partially addresses #3, as with MakD v1.6.0 documentation
can be generated by using `make doc`.