Break up the line used to format an array
In the 'serializeArray()' function, break the single format call used to
format an array into multiple statements. This is done in order to
completely avoid calling format for null arrays.

When moving to using the new 'ocean.text.convert.Formatter' module,
formatting null arrays results in the string '{null}' instead of an
empty string.