Move SignalExt from SignalFD to regular signal dispatch mechanism
Using `signalfd` has certain disadvantages when used to wait for the
signals, which all originates from the point that the signals must be
blocked for delivery all the time, and that normal signal disposition
mechanism is not used (for example:
https://bugzilla.kernel.org/show_bug.cgi?id=9039)

This patch stops using signalfd in SignalExt. Instead it uses the
regular signal disposition mechanism, used with `pipe (2)` for queuing
the signals while waiting for the free epoll cycle which will process
these events.

Fixes #1459