travis: Revamp docker handling and push to registry
A few assorted improvements to travis scripts. First split into separated
scripts the before_install and script commands. Then use the new
sociomantictsunami Docker Hub organization. Finally try to pull the image
from the registry, so we can use it as a cache, and try to push it back
if there are credentials available.

Also use $TRAVIS_REPO_SLUG to get the docker image name, so scripts are
more reusable.