Fix concatenated gzip support for ZlibStream
The existing design of ZlibStream assumed that Z_STREAM_END means EOF, and
used this to automatically deallocate resources. But, gzip files can be
concatenated. In this case, Z_STREAM_END may be followed by additional data
chunks. This means that it is impossible for ZlibStream to know if more
data is coming, so it cannot perform automatic resource management.

To fix this, the return value of decodeChunk() is removed, and instead, a
new end() function is added, which should be called after the last call
to decodeChunk().