docker: Move scripts used by CI to ci
Mixing stuff run by docker's image creation and stuff that run inside
the container is a bad idea, because any changes in the files copied to
the docker daemon, can potentially invalidate the image cache triggering
an unnecessary rebuild.

Because of this is better to have scripts used to build the docker image
separated from the scripts used to actually run the tests in a CI
system.

This commit also moves the Dockerfile to the top-level directory to make
it easier to just run `docker build .` (for which a `.dockerignore` file
is added, so only stuff in `docker/` is copied to the docker daemon).