Fork client process after listen() in tests
Previously, the client process was created before the server process
was guaranteed to call bind/listen. Because of this, the tests were
flaky, since the connect() could be called before bind()/listen()
was done. This forks after bind/listen (which is necessary for
setting a backlog on a socket), so the connect() is guaranteed to be
ready to be called.

Note that, although forked process in this case inherits server socket,
unless it `accept`'s, no connect() attempts will be redirected to it.