Fix StopWatch documentation for integer vs. floating-point times
The assertion that floating-point imposes "no implicit granularity to the
measurable intervals, except the limits of fp representation" is deceptive:
it is true only for small values, the limits of fp representation impose a
more coarse granularity for greater values. The correction also mentions
the advantage of absolute integer precision which always yield the expected
result when comparing.