Format Time correctly with Formatter
Before a Time structure was formatted as any struct (printing the only member, ticks_).
With this change we get a behaviour closer to the one of Tango_Layout, as it formats it as a human-readable string.
However, the formatting is currently limited to the 'G' formatting, as we lack a way to transmit format specifier to 'toString'