Pass SmartUnion to FileSystemEvent's callbacks.
`FileSystemEvent`'s constructor and `setNotifier` now accepts
`FileSystemEvent.Notifier` which accepts the SmartUnion of the possible
events that can be generated
(with one union member carrying the name as well as the path).

This is useful for monitoring the directories, where the name of the affected
file will be passed to this handler.

In addition, the old Handler is deprecated.