Iterate over all pool items in awaitRunningTasks
In the 'awaitRunningTasks()' function, instead of iterating over only
the busy tasks in the task pool, iterate over all tasks instead. But
continue the iteration loop immediately if the task being iterated upon
is discovered to be busy.

This change retains the exact same semantics, but will allow us to
perform other checks in future.