AddressIPSocket: Fix EFAULT on invalid IP address presentation string
Make AddressIPSocket methods fail with EAFNOSUPPORT as their documentation
specifies, not EFAULT.

IP address presentation strings (that is, strings that are expected to
contain the address in the "a.b.c.d" notation) accepted by AddressIPSocket
methods are forwarded to IPAddress.opCall(), which calls inet_pton(3) to
parse them, and returns a sockaddr* pointer on success or null on failure,
i.e. inet_pton failed. On failure the returned null pointer was passed to
connect(2), which then fails with EFAULT "Bad address" (this would be a
segfault in user space).

Ironically "Bad address" is actually a helpful error message to report an
invalid IP address.