Fix documentation for ocean.text.convert.Formatter.sformat
The original documentation stated that the formatted string would be
written into the provided buffer, when in fact it is appended to it.

The note about the buffer being 'extended if needed' has been replaced
with a note that the capacity will be increased if needed, since we can
reasonably assume that appending will always extend the buffer unless
the format string `fmt` is empty.