travis: Build matrix for F,DC,DIST combinations
This commit configures Travis to use matrix builds for all the
combinations of F={production,devel} (flavours),
DC={dmd1,dmd-transitional} (compiler, in the future upstream DMD could
be added as a "allow_failures" combination) and DIST={trusty,xenial}
(Ubuntu version).

For now the Dockerfile is sort of duplicated, but since it's minimal, it
shouldn't be a problem anyway.