Deprecate contiguous serializer traits
Those originated in over-generalized assumption that new framework will
be use to adjust all over serializers but it proved to be impactical for
the time.

At the same time extra compile-time reflection complicated code
structure consierably and resulted in suppression of relevant compiler
messages. It is better to deprecate/remove it for now and reconsider if
standardization of serializer is to happen again.