Formatter: Do not explicitly instantiate the non-deprecated sformat
This triggers an ICE in DMD1:
dmd1: template.c:4644: Identifier* TemplateInstance::genIdent(): Assertion 'i + 1 == args->dim' failed.
It happens because some code in swarm is calling the deprecated overload of sformat,
and the ICE is triggered without showing any deprecation message. It does not happen in D2.
Attempts to craft a test case to reproduce the issue were unsuccessful.