Deprecate old loggers
This patch is a bit more complicated than it ought to be.
As ocean accepts no deprecation, we cannot call `configureOldLoggers` from `LogExt`,
however we do want to keep configuring them, while deprecating the function.
This is why an `extern (C)` hack (and 2 alias, to work around bugs) are introduced.

In addition, deprecating the `Logger` type triggers deprecations when `HierarchyT`
is instantiated with it as argument, even if it's through a `deprecated alias`
or even if it's a dummy `deprecated class` inheriting from it.

So instead of deprecating the `Logger` type, we make it totally useless by deprecating
most of its methods, as well as the mean to look it up (the `Log` structure).

We couldn't deprecate the module, as it needs to be imported sometimes.
As we cannot import from within a function (D1 limitation), the original solution
was to import from a deprecated aggregate (e.g. struct) and use aliases from there.
Sadly this still triggers deprecations with our DMD transitional.