Replace dodgy `debug (StopWatch)` code with a viable unittest
This debug check appears to have been written in the deep dark days of
prehistory: it is not even compatible any more with current unittesting
practice.

This patch implements a unittest replacement for the `StopWatch` struct,
using `nanosleep` rather than an unreliable [1] `for` loop in order to
ensure the delay that the `StopWatch` requires.  As a side effect, this
unittest demonstrates that the documentation of the `stop` method is
telling us lies: it does not `stop` the timer at all, merely returns the
current accumulated time in seconds (akin to what physical stopwatches
would call the `split`).

  [1] (the fear here is that a decent optimizing compiler could well
       entirely remove the `for` loop, since it does nothing...)

As an amusing observation, it seems that the actual count of time that
the stopwatch records is rather greater than the 10 microseconds that
`nanosleep` requests.  It looks like that 'at least the time specified'
remark in the `nanosleep` documentation is being taken at significantly
more than face value ... :-)