Deprecate `StopWatch.stop` in favour of `StopWatch.sec`
The old `stop` method of `StopWatch` was both misleadingly named and
misleadingly documented: contrary to what it claimed, it did not stop
the timer.  With this in mind it has been renamed to `sec` (by analogy
to the existing `microsec` method).

A deprecated alias has been provided to allow the old name to continue
to be used, and a nearly-duplicate unittest has been added so both old
and new usages will continue to be tested.