Add BTree and operations over it to ocean
This commits introduces BTree data structure and algorithms over it to
the ocean. This implementation of a BTree is templated on the type it
carries, on the tree's degree, and on the allocation policy that should
be used for the allocating the internal nodes (it is useful if the
extending the BTree is done once in a while, but when we shouldn't give
chance to GC to do mark&sweep). Additionally, GC tracking of the memory
can be turned of, so the GC will not (somewhat expensively) traverse the
tree, which is useful if we're storing value types, or the reference
types for which we're storing the reference somewhere else.

In addition to basic functionality, inorder traversal of the tree,
either via opApply or InputRange is supported.