Remove redundant scope around and else branch for enforces
Right after an if statement an enforce is placed inside its own scope.
This is a leftover from old code. The intention here is to always
enforce. For clarity the scope is removed since otherwise one might
think an else branch is missing.
In a related case, a few lines above, the enforce is placed inside an
else branch. But here again it was intended to always perform the
enforce for clarity and ease of refactoring.

This change is not bug fix. It is merely changing the code to be easier
to understand and maintain.