Don't unregister the TimerSet and timers if exception is thrown in TimerExt
`TimerExt` is now swallowing all the unhandled exceptions, keeping the timers
registered by default. Previously, the uncaught exception would
propagate to TimerSet, unregistering it and making all TimerExt
registration not work again. That behaviour is now fixed, with the
default that, if exception is thrown, the timer will stay registered.