Don't decrement length of cache before removing item
`replaceRemovedItem` switches the two items referred to by its
two arguments. This is used when removing an item by copying the
last item in the list over the item to be removed, then
decrementing the length of the cache. The `insert` field represents
the length of the cache. `replaceRemovedItem` calls `length` -- which
returns the value of `insert` -- to assert that the indices of the
items to be swapped are in range. Therefore, it is necessary to
decrement `insert` (==`length`) only *after* having called
`replaceRemovedItem`.