XML parser should copy the name and value rather than using a slice
The XML parser would previously rely on slices to the names and values
for the nodes. This resulted in memory being overwritten when multiple
nodes were added with values from the same buffer.

To fix this issue, the names and values should be copied to the node
arrays rather than relying on the slices. To prevent memory allocation
issues, the reset method has been modified to set the length of the
arrays to 0 rather than null.