Deprecate thin wrappers in ocean.core and VersionIdentifiers
Those thin wrappers are not useful and just create an extra indirection.
VersionIdentifiers date from the time Makd was part of ocean, which is long gone.
In the process of deprecating `ocean.core.Vararg`, several useless `version` blocks were cleared,
as ocean is only compiled with DMD on X86_64 machines, and those are essentially dead code.