Allow setting mode for the socket in UnixSocketExt
`UnixSocketExt` now supports setting the mode for the socket after
binding. User can set it in the config file in `UNIX_SOCKET.mode`
as an octal string and the mode will be applied to the socket after
creating it. This allows setting mode which would possibly prevent
all other users for connecting to the socket (most likely choice would
be `0660`).