Support reopening single files via unix domain socket in DaemonApp
`DaemonApp`'s `ReopenableFilesExt` now supports reopening commands from the
UnixSocketExt. If you have the UnixSocketExt bound to the socket, sending command
`reopen_logfile file_path` will reopen this file, if registered, and acknowledge
the action with `ACK` or `ERROR` strings.

This is different than reopening files with the reopen signal, which
is not capable passing the file path to it, so it forces reopening all
files.

Fixes #190