Add D2 auto-conversion for tags
Now when a tag is released, a new stage will also convert the tag to D2,
then it will commit the changes, create a new tag with a +d2 suffix,
push the changes and create a GitHub release for it.

This is done via the new beaver dlang d2-release command.

Also update beaver to v0.3.0 to use its new features.

* submodules/beaver v0.2.2(480e6f5)...v0.3.0(3f00ea2) (14 commits)
  > Convert release notes to the new format
  > dlang make: Upload codecov reports automatically
  > dlang: Add command to send reports to codecov
  > dlang: Pass a few extra Makd env variables to make
  > dlang: Add a command to release converted D2 tags
  (...)