Unregister connection socket before closing it
As noted in https://github.com/sociomantic-tsunami/dhtnode/issues/25#issuecomment-339293270,
it is unsafe (at least in D1 builds with the CDGC) to rely on the OS to
unregister a socket from epoll when it's closed. When the GC kicks in, it
forks, causing all fds of the process to be duplicated. While the fork is
alive, this means that closing a socket that existed when the GC fork began
will *not* cause it to be unregistered from epoll, as the fork's copy
remains. This can result in events firing for a socket that has been
closed.

This commit makes sure that a closed connection socket cannot fire in epoll
by unregistering it before closing it.

Note that TaskConnectionHandler will also need to be adapted to override
unregisterSocket(), but this will likely require new features and is not
regarded as an urgent bugfix (as TaskConnectionHandler is not yet in active
use).