Fix the broken D2 conversion of PullParser
PullParser was converted to D2 incorrectly. Internally, the code does not
modify the text which is being parsed. But the D2 conversion incorrectly used
an `mstring` to store the text being parsed. This required a memory
allocation for every cstring which was parsed, making the code essentially
unusable in D2.

The `name` and `value` member functions were incorrectly marked as returning
mutable strings. To entirely avoid breaking changes, that is is not fixed in
this PR, but the documentation now states that they may allocate memory.