Document.Node: Don't store mutable slices in nodes
The NodeImpl class is problematic for D2, because it contains
slices of the text being parsed (which is conceptually const) but uses those
same slices as reusable buffers (which are mutable).
To fix this, introduce a new ConcatBuffer member to Document to store the
modifierd text, while changing the slices to be const.