Fix `drop` not working after `unregister`
Follow-up to b084f6b347fdc9495c74d600ed88924c648e274d which fixes the
problem with expiry reference being cleaned up on call to `unregister`
which is done when timeout hits. That resulted in not being able to
`drop` it later, which killed the very point `drop` was added.