Fix ocean.io.Path.isWritable
The previous implementation of isWritable was broken, as it was using
stat function inappropriately, combining it with a O_RDONLY flag from
the open function, which also happens to has value of 0 so it would
always report is writable. On top of this, even if it worked by some
accident, it would still completely ignore the effective/real uid and
the file ownership.
Although now used access doesn't check the real uid/gid, but the
effective one (meaning the succeeding open could still fail/pass), it's
what we actually want most of the time.