Add module to shallow reset a struct
The main motivation for introducing this functionality (and not simply
using the existing DeepReset functionality) is that DeepReset recurses
into and individually resets every element of each dynamic array present
in the struct. However, this additional work is sometimes overkill for
certain performance critical applications where just resetting the
lengths of the dynamic arrays to zero suffices.