Upgrade to supported makd version
Removing Config.mak, per Leandro explanation:

Build.mak, if you need to use other variables normally you want to do it
in Build.mak. Config.mak is only for setting defaults not depending on
other variables, as there is almost nothing defined when Config.mak is
processed. Config.local.mak is the same but for local configuration that
you don't want to check in, for example if you like to compile always in
verbose mode, set V :=1 in Config.local.mak.