Fix 429: No error message when command line argument has error
The regression appeared in #362, specifically in commit 480fb9fcd71a89e8e93fe917937fda258b08a5aa.
The change in this commit was wrong, it used `sprint` which is equivalent to `snformat`,
(which write to a buffer up to the buffer's size,
when all we really wanted was something that appended to the buffer.