Fix release notes
A mixed combination of old and new formatting is used in release notes,
not only in different files but even also in the same file. There are
also a couple of other formatting issues and some other minor things to
fix.

This commit should make everything more consistent and clean.