Unify Appender.Layout.format delegate type with the rest of Ocean
Since `size_t` delegate were little or not used (for various, valid reasons),
we moved away from them.
Moreover, there's no point in accepting a `Const!(void)[]` if the data is casted under the hood,
as it essentially relies on the data being `cstring` without the safety of type checks.
For those reasons, the delegate types were updates.
This is a breaking change which should have little impact on downstream, if any.