Improve Exceptions in ocean.sys.Process
The ctors had two issues: the main one, they were not forwarding their file and line arguments, so it was lost.
The second one, they did not provide default __FILE__ and __LINE__ default arguments as is common with exceptions ctor,
leading to rather non-DRY code.
The two ctors of ProcessCreateException have been merged, as a call to new ProcessCreateException(istring, istring) was ambiguous
after the addition of default parameters.