Allow removing the client from the selected set in EpollDispatcher.unregister
`EpollSelectDispatcher.unregister` now accepts an optional flag
(`remove_from_selected_set`) which will remove the client from the
selected set if it's in there. This will allow other ISelectClients to
safely destroy other ISelectClients from their handle method
(previously, this was unsafe because even unregistering the client
would just remove it from the all next epoll cycles, but not from the
other one, if the client's fd reported that it's ready to be handled).