BufferedInput: Extend the buffer if needed
The iterator support in BufferedInput was quite impractical:
The user needed to know, on instantiation, the maximum size of a "record" in case the iterator interface is used.
In practice, it is used by stdin to read line by line, however one cannot know how large a line is.
Extending the buffer, fixes this issue.