Don't schedule the timer interval to that very moment
Due to the application's misdesign, it might happen
that not all resources that are used in `onStatsTimer` are
initialized at the point when the `startEventHandling` is
called. This leads then to the issue when the Application
starts at the time which would be the aligned time point
and there's no wait between next onStatsTimer call.

This ensures that there's at least one second of the
delay between onStatsTimer and startEventHandling.