Add thread workers support
This adds support for calling user delegates in a task-blocking fashion
which will be then executed in separate threads. This makes it possible
to perform blocking system calls (reading from disk or, for example,
using curl_easy_* API) as a task-blocking operations.

For usage example, see integrationtests/asyncio/main.d.