Ignore tasks scheduled during shutdown
Because of missing return statement, trying to schedule a new task
after shutdown would succeed but result in application infinitely stuck
during shutdown sequence.

With this fix, any such tasks will be ignored allowing for clean exit.
There is a concern of losing app data but it will have to be addressed
in a separate changest, targetting minor or maybe even major branch.