Allow for dynamic terminal width inside AppStatus
AppStatus was assuming that main terminal exists and that
it can be queried for its width. However this is not true
when starting programs as daemons from systemd - the main terminal
doesn't exist (stdout may or may be not redirected), so the
width is zero. However, user may want to attach to the unix socket
and get the output from the AppStatus.

AppStatus is changed so it a) accept assumed terminal width at
construction which, b) can be overridden by passing the width
to the command (`app_status 120` for example).