Support copying empty Contiguous instances
The original implementation of `ocean.serialize.contiguous.Util.copy`
will throw a `SanityException` if `src` contains no data, because the
`Deserializer.deserialize` method requires that underlying struct data
is present.  This is non-intuitive, because there is no reason why it
should be forbidden to deep-copy empty instances: the natural outcome is
that the `dst` parameter then also winds up empty.

This patch fixes the issue by ensuring `copy` checks `src.ptr` and then
if it is `null` resets `dst` instead of deep-copying to it.

Note that we need to temporarily cast away `const`-ness of `src` to work
around the fact that the `ptr` method does not currently support it.  We
do not modify any underlying data, so this should not break the promise
of the `copy` function signature.

Fixes https://github.com/sociomantic-tsunami/ocean/issues/652.