Ensure closures script fails if make fails
The intent behind this version of the script is that it should fail if
`make` fails, if `grep` fails, or if `grep` finds a matching line; in
other words, if the exit status of `make` is non-zero or that of `grep`
is not 1 (no lines found).

The `set -o pipeline` statement in the original script means non-zero
exit statuses from `make` propagate to the exit status of `grep`, but
the leading `!` means this is inverted to result in a false test pass.

Applying the `!` only to the output status of `grep` itself would fix
this, but leaves a second problem: `grep` may itself fail, in which case
it will report exit status 2.

This patch therefore rewrites the check to use `test` to ensure that the
exit status of `grep` is 1.  This means we can no longer use `set -e` as
this would mean that the `grep` command would fail: instead we carry out
manual checks on the exit status of each command.

In order to simplify the script, we replace the `make | grep` pipe with
a multi-line version where `make` output is written to a tempfile.  This
allows us to drop `set -o pipeline`.

This should mean that the script will cause a test failure in any of the
following circumstances:

  * `make` fails (exit status non-zero) for either of the 2 calls
  * `grep` fails (exit status 2)
  * `grep` finds at least one line containing `closure` (exit status 0)

... which in all cases will result in the script exiting with status 1.
If on the other hand `make` passes and `grep` does not find `closure`,
we will get the zero exit status required to pass the test.

Note that this update may mean that the closure script will now fail for
`+d2` auto-converted releases: the hardcoded `make d2conv` means that it
will run `d1to2fix` on already-converted code, and thus `make fasttest`
will fail.