Add select cycle callbacks
Select cycle callbacks are delegates that are
executed once the select cycle has finished.

These "one-shot" callbacks are stored in a queue,
and have to be registered each time from new.

Note that the select cycle (i.e. `epoll_wait`) doesn't
block as long as the callback queue is not empty.