EmailSender: Fix the envelope sender address
Set the envelope sender address as it was empty and
this is where delivery problems are sent to.
Merely setting `From` in the header will not be
sufficient and bounces will return to the wrong place.

Also if `From` in the header does not match the envelope
sender, the e-mail could be considered spoofing and gets
rejected.