Workaround a compiler bug
We had reports of this code issuing a compile error with dmd 1.081.2 and
1.082.1 when -release -O -inline is used. The errors being:

./submodules/ocean/src/ocean/util/container/cache/ExpiringCache.d(346): Error: null dereference in function _D5ocean4util9container5cache13ExpiringCache22__T13ExpiringCacheVm8Z13ExpiringCache19getExpiringOrCreateMFmJbbZAv
./submodules/ocean/src/ocean/util/container/cache/ExpiringCache.d(355): Error: null dereference in function _D5ocean4util9container5cache13ExpiringCache22__T13ExpiringCacheVm8Z13ExpiringCache19getExpiringOrCreateMFmJbbZAv
./submodules/ocean/src/ocean/util/container/cache/ExpiringCache.d(362): Error: null dereference in function _D5ocean4util9container5cache13ExpiringCache22__T13ExpiringCacheVm8Z13ExpiringCache19getExpiringOrCreateMFmJbbZAv

Looking at those lines of code (in v4.4.0+breaking) is quite clear that
cache_item can't really be `null`, there is even an `assert()` checking
for that, and it would be impossible for the compile it to determine it
at compile time, so this really looks like a compiler bug.

For now we just convert the assert() to verify() to make sure the check
is not removed by -release, and the optimizer doesn't get confused.