Refactor ThrottledTaskPool to remove overrides of start()/restore()
These methods were duplicating the existing methods but adding method calls
for triggering suspension. this was in the past because provided tasks
were overwritten as the task system didn't support termination hooks.

Now that the derived internal class is removed these methods can be
refactored away and the logic can be done in an override of startImpl.