Support extra args in task pool deserialization.
When restoring disk, there was no way to restore runtime information
that is unable to be written to disk (eg. objects, function pointers).
In the past the suggested method was to have a static set method
applicable to all tasks but this change allows the user to pass additional arguments to the
deserialize method.