Replace getIeeeFlags with version in phobos std/math
ocean almost compiles with gdc, if it weren't for this function and
a few other small things. Sadly I have no idea, what this function does,
but a function with the same name exists in phobos, which in contrast to
the ocean one compiles with both dmd and gdc, so it seems cool to me.