Replace D1 unary operator overloads in ocean.math
DMD 2.088.0 has deprecated D1 operator overloads, which are expected to
be replaced with the newer D2 `op{Unary,Binary,BinaryRight,OpAssign}`
templated methods: https://dlang.org/changelog/2.088.0.html#dep_d1_ops

`ocean.math` contains surprisingly few such operators, and they are easy
to replace.  For `WideUInt` there should be no resulting problems, as it
is a struct, but `Distribution` and `SlidingAverageTime` are classes, so
there is a risk this will be a breaking change for downstream subclasses
if any attempt to override the older D1 methods.