Deprecate D1 operator overloads in SignalSet struct
DMD 2.088.0 has deprecated D1 operator overloads, which are expected to
be replaced with the newer D2 `op{Unary,Binary,BinaryRight,OpAssign}`
templated methods: https://dlang.org/changelog/2.088.0.html#dep_d1_ops

The `SignalSet` struct has a slightly unusual setup, using custom `add`
and `remove` methods doing the real work, and using aliases to these to
define addition and subtraction operator overloads.

Since this is a struct rather than a class, we do not have to care about
whether these methods are overrideable.  What is more problematic is the
fact that as written, these aliases do not correctly implement addition
or subtraction operators, which are supposed to return the result.

It therefore looks like `b + c` will update `b` in place and result in a
`void` result (i.e. we cannot use the result of the addition).

This honestly feels like a serious code smell, and in any case it seems
inappropriate to use `+` and `-` operations to represent insertion and
removal from a set.  Within ocean itself, `SignalMask` is only ever used
in `ocean.sys.SignalFD`, where the `add` and `remove` methods are used,
never the operators.  I think it likely that the operator overloads were
never actually used, but perhaps included because it seemed a good idea
to someone at the time, and then forgotten about.

Given all these factors, rather than replace these operators with their
D2 equivalents, the existing operators have just been deprecated as-is,
which should alert any downstreams relying on this weird behaviour, and
paves the way for removal in the next major release.

Part of https://github.com/sociomantic-tsunami/ocean/issues/747.