Replace Const! with const
In D2 we can use `const` and don't have to use the `Const` template.
This change replaces all `Const!(` with `const(` using
`'s/\(^\|\s\+\|(\)Const!(/\1const(/g'` with the sed tool on the code
base including integration tests. This means it will only replace it,
when `Const!(` is at the beginning, has any preceding white space or an
preceding opening brace. This covers all cases when `Const!(` is used.
More importantly, it excludes usages of the template
`ocean.io.digest.Fnv1.Fnv1Const`.

Note, that the above also includes the unittest for the `Const`
template. But it was left out to have the unittest for `Const`.