Replace Immut! with immutable
In D2 we can use `immutable` and don't have to use the `Immut` template.
This change replaces all `Immut!(` with `immutable(` using
`'s/\(^\|\s\+\|(\|\[\)Immut!(/\1immutable(/g'` with the sed tool on the code
base including integration tests. This covers all uses uses of `Immut!` in the
code base. But this change manually excluded the unittest for `Immut` to
have it unchanged.