Replace Inout! with inout
In D2 we can use `inout` and don't have to use the `Inout` template.
This change replaces all `Inout!(` with `inout(` using
`'s/\(^\|\s\+\|(\|\[\)Inout!(/\1inout(/g'` on the code base including
integration tests. This covers all uses uses of `Inout!` in the code
base. But the unittest for `Inout` is kept as is to have it test the
`Inout` template.