Add support for Ubuntu bionic
Set libssl and libcryto to version 1.0.0 supported in Ubuntu xenial and bionic.
The specific version will have to be removed once the support for xenial is
dropped and the bindings to libssl have been updated to v1.1.x which is
supported by default in bionic.

The dmd-transitional compiler is available for Ubuntu bionic only in
version 2.078.3.s*.